fix controller

+++ app/Http/Controllers/Backsite/DoctorController.php
@@ -6,8 +6,8 @@ use App\Http\Controllers\Controller;
 
 // use library here
 use Illuminate\Support\Facades\Storage;
-//use Illuminate\Support\Str;
-//use Illuminate\Http\Request;
+use Illuminate\Support\Str;
+use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
 // request
@@ -15,12 +15,12 @@ use App\Http\Requests\Doctor\StoreDoctorRequest;
 use App\Http\Requests\Doctor\UpdateDoctorRequest;
 
 // use everything here
-//use Gate;
+use Gate;
 use Auth;
-//use File;
+use File;
 
 // use model here
-//use App\Models\User;
+use App\Models\User;
 use App\Models\Operational\Doctor;
 use App\Models\MasterData\Specialist;
 
@@ -42,13 +42,18 @@ class DoctorController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('doctor_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         // for table grid
         $doctor = Doctor::orderBy('created_at', 'desc')->get();
 
         // for select2 = ascending a to z
         $specialist = Specialist::orderBy('name', 'asc')->get();
+        $user = User::whereHas('detail_user', function($query){
+                    $query->where('type_user_id', 2);
+                })->orderBy('name', 'asc')->get();
 
-        return view('pages.backsite.operational.doctor.index', compact('doctor'));
+        return view('pages.backsite.operational.doctor.index', compact('doctor', 'specialist', 'user'));
 
     }
 
@@ -72,6 +77,25 @@ class DoctorController extends Controller
     {
         // get all request from frontsite
         $data = $request->all();
+
+        // re format before push to table
+        $data['fee'] = str_replace(',', '', $data['fee']);
+        $data['fee'] = str_replace('IDR ', '', $data['fee']);
+
+        // upload process here
+        $path = public_path('app/public/assets/file-doctor');
+        if(!File::isDirectory($path)){
+            $response = Storage::makeDirectory('public/assets/file-doctor');
+        }
+
+        // change file locations
+        if(isset($data['photo'])){
+            $data['photo'] = $request->file('photo')->store(
+                'assets/file-doctor', 'public'
+            );
+        }else{
+            $data['photo'] = "";
+        }
 
         // store to database
         $doctor = Doctor::create($data);
@@ -99,10 +123,15 @@ class DoctorController extends Controller
      */
     public function edit(Doctor $doctor)
     {
+        abort_if(Gate::denies('doctor_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         // for select2 = ascending a to z
         $specialist = Specialist::orderBy('name', 'asc')->get();
+        $user = User::whereHas('detail_user', function($query){
+                    $query->where('type_user_id', 2);
+                })->orderBy('name', 'asc')->get();
 
-        return view('pages.backsite.operational.doctor.edit', compact('doctor', 'specialist'));
+        return view('pages.backsite.operational.doctor.edit', compact('doctor', 'specialist', 'user'));
     }
 
     /**
@@ -116,6 +145,32 @@ class DoctorController extends Controller
     {
         // get all request from frontsite
         $data = $request->all();
+
+        // re format before push to table
+        $data['fee'] = str_replace(',', '', $data['fee']);
+        $data['fee'] = str_replace('IDR ', '', $data['fee']);
+
+        // upload process here
+        // change format photo
+        if(isset($data['photo'])){
+
+            // first checking old photo to delete from storage
+           $get_item = $doctor['photo'];
+
+           // change file locations
+           $data['photo'] = $request->file('photo')->store(
+               'assets/file-doctor', 'public'
+           );
+
+           // delete old photo from storage
+           $data_old = 'storage/'.$get_item;
+           if (File::exists($data_old)) {
+               File::delete($data_old);
+           }else{
+               File::delete('storage/app/public/'.$get_item);
+           }
+
+       }
 
         // update to database
         $doctor->update($data);
@@ -132,6 +187,18 @@ class DoctorController extends Controller
      */
     public function destroy(Doctor $doctor)
     {
+        abort_if(Gate::denies('doctor_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
+        // first checking old file to delete from storage
+        $get_item = $doctor['photo'];
+
+        $data = 'storage/'.$get_item;
+        if (File::exists($data)) {
+            File::delete($data);
+        }else{
+            File::delete('storage/app/public/'.$get_item);
+        }
+        
         $doctor->forceDelete();
 
         alert()->success('Success Message', 'Successfully deleted doctor');
